requirements.txt
anyio==3.6.2
bcc==0.18.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.2.0
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
h11==0.12.0
hidpidaemon==18.4.6
httpcore==0.15.0
httplib2==0.20.2
httpx==0.23.0
idna==3.3
importlib-metadata==4.6.4
install==1.3.5
jeepney==0.7.1
kernelstub==3.1.4
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
more-itertools==8.10.0
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.0
pop-transition==1.1.2
protobuf==3.12.4
pycairo==1.20.1
pycups==2.0.1
pydbus==0.6.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyTelegramBotAPI==4.7.0
python-apt==2.3.0+ubuntu2.1
python-debian===0.1.43ubuntu1
python-gnupg==0.4.8
python-xlib==0.29
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.27
PyYAML==5.4.1
repolib==2.0.0
repoman==1.4.0
requests==2.25.1
rfc3986==1.5.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
sessioninstaller==0.0.0
six==1.16.0
sniffio==1.3.0
systemd-python==234
tornado==6.2
tzdata==2022.5
tzlocal==4.2
ubuntu-advantage-tools==27.11.2
ubuntu-drivers-common==0.0.0
ufw==0.36.1
urllib3==1.26.5
wadllib==1.3.6
xdg==5
xkit==0.0.0
zipp==1.0.0
